create historical prices repository and store dummy historical price
