+++ src/main/java/org/galatea/starter/domain/IexHistoricalPrices.java
@@ -1,20 +1,2 @@
-package org.galatea.starter.domain;
-
-import java.math.BigDecimal;
-import java.util.Date;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class IexHistoricalPrices {
-
-  private BigDecimal close;
-  private BigDecimal high;
-  private BigDecimal low;
-  private BigDecimal open;
-  private String symbol;
-  private Integer volume;
-  private String date;
-
+package org.galatea.starter.domain;public class IexHistoricalPrices {
 }
